Move expressions to their own folder.
