+++ src/CoreEx.Data/Querying/QueryFilterParser.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
 
+using CoreEx.Data.Querying.Expressions;
 using CoreEx.RefData;
 using System;
 using System.Collections.Generic;
